LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.22-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Expected an identifier but found '=' instead
S7.5.2_A1.22

PASS Expected parsing failure

TEST COMPLETE
